requirements.txt
# Core dependencies
sqlalchemy>=2.0.0
alembic>=1.13.0

# Document parsing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0
markdown>=3.5.0
ebooklib>=0.18
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Audio processing
pydub>=0.25.1
mutagen>=1.47.0

# Text processing
nltk>=3.8.0
regex>=2023.12.0

# CLI & utilities
click>=8.1.0
rich>=13.7.0
pyyaml>=6.0.3
python-dotenv>=1.0.0

# Validation
pydantic>=2.5.0

# Progress tracking
tqdm>=4.66.0
